+++ client/client_settings.hpp
@@ -58,6 +58,7 @@ struct ClientSettings
     struct Player
     {
         std::string player_name{""};
+        std::string parameter{""};
         int latency{0};
         PcmDevice pcm_device;
         SampleFormat sample_format;