+++ client/player/file_player.cpp
@@ -22,6 +22,7 @@
 #include "common/aixlog.hpp"
 #include "common/snap_exception.hpp"
 #include "common/str_compat.hpp"
+#include "common/utils/string_utils.hpp"
 #include "file_player.hpp"
 
 using namespace std;
@@ -31,8 +32,33 @@ static constexpr auto kDefaultBuffer = 50ms;
 
 
 FilePlayer::FilePlayer(boost::asio::io_context& io_context, const ClientSettings::Player& settings, std::shared_ptr<Stream> stream)
-    : Player(io_context, settings, stream), timer_(io_context)
+    : Player(io_context, settings, stream), timer_(io_context), file_(nullptr)
 {
+    auto params = utils::string::split_pairs(settings.parameter, ',', '=');
+    string filename;
+    if (params.find("filename") != params.end())
+        filename = params["filename"];
+
+    if (filename.empty() || (filename == "stdout"))
+    {
+        file_.reset(stdout, [](auto p) { std::ignore = p; });
+    }
+    else if (filename == "stderr")
+    {
+        file_.reset(stderr, [](auto p) { std::ignore = p; });
+    }
+    else
+    {
+        std::string mode = "w";
+        if (params.find("mode") != params.end())
+            mode = params["mode"];
+        if ((mode != "w") && (mode != "a"))
+            throw SnapException("Mode must be w (write) or a (append)");
+        mode += "b";
+        file_.reset(fopen(filename.c_str(), mode.c_str()), [](auto p) { fclose(p); });
+        if (!file_)
+            throw SnapException("Error opening file: '" + filename + "', error: " + cpt::to_string(errno));
+    }
 }
 
 
@@ -56,7 +82,7 @@ void FilePlayer::requestAudio()
     if (buffer_.size() < needed)
         buffer_.resize(needed);
 
-    if (!stream_->getPlayerChunk(buffer_.data(), 100ms, numFrames))
+    if (!stream_->getPlayerChunk(buffer_.data(), 10ms, numFrames))
     {
         // LOG(INFO, LOG_TAG) << "Failed to get chunk. Playing silence.\n";
         memset(buffer_.data(), 0, needed);
@@ -65,7 +91,8 @@ void FilePlayer::requestAudio()
     {
         adjustVolume(static_cast<char*>(buffer_.data()), numFrames);
     }
-    fwrite(buffer_.data(), 1, needed, stdout);
+    fwrite(buffer_.data(), 1, needed, file_.get());
+    fflush(file_.get());
     loop();
 }